1. remove Handle32 and Handle64, using HANDLE. 2. Succeed to run 64bit Elastos Applications on Pixel with Android O.

+++ Sources/Elastos/Frameworks/Droid/Base/Core/inc/elastos/droid/os/BinderProxy.h
@@ -45,7 +45,7 @@ private:
     CARAPI_(void) Destroy();
 
 public:
-    Int32 mObject;
+    Int64 mObject;
 };
 
 } // namespace Os
