+++ Sources/Elastos/Runtime/Core/reflection/CInterfaceInfo.h
@@ -122,7 +122,7 @@ private:
     UInt32  mIndex;
     UInt32  mIFCount;
     UInt32  mMethodCount;
-    Int32   mBase;
+    uintptr_t  mBase;
 };
 
 #endif // __CINTERFACEINFO_H__
